28.02.19 --> SQLParser comming soon

+++ MiniSQLEngine/SQLParser.cs
@@ -23,20 +23,38 @@ namespace MiniSQLEngine
             string select5 = @"SELECT\s+(\w+)(\,\s+(\w+))+\s+FROM\s+(\w+);";
             string select6 = @"SELECT\s+(\w+)(\,\s+(\w+))+\s+FROM\s+(\w+)\s+WHERE\s+(\w+)\s*(=|<|>)\s*(\w+);";
 
-            string insert = @"(INSERT)\s+(INTO)\s+(\w+)(\s+|\s+\((\w+)\)\s+|\s+\((\w+)(\,\s+(\w+))+\))\s+(VALUES)(\s+\((\w+)\)|\s+\((\w+)(\,\s+(\w+))+\));";
-            string delete = @"(DELETE)\s + (FROM)\s + (\w +)\s + (WHERE)\s +\w +\s * (=|<|>)\s *\w +;";
-            string update = @"(UPDATE)\s+(\w+)\s+(SET)\s+(\w+)\s*\=\s*(\w+)(\s+|\,\s+(\w+)\s*\=\s*(\w+)\s+)(WHERE)\s+(\w+)\s*(=|<|>)\s*(\w+);";
+            //Insert
+            string insert1 = @"INSERT\s+INTO\s+(\w+)\s+VALUES\((\w+)\);"; //CON TODOS SUS VALUES(1)
+            string insert2 = @"INSERT\s + INTO\s + (\w +)\s + VALUES\s +\(*(\w +)(\,\s + (\w +))+\);"; //(CON TODOS SUS VALUES(+1))
+            string insert3 = @"INSERT\s + INTO\s + (\w +)\s +\((\w +)\)\s + VALUES\s +\((\w +)\);"; //(CON UN VALUE)
+            string insert4 = @"INSERT\s + INTO\s + (\w +)\s +\((\w +)(\,\s + (\w +))+\)\s + VALUES\s +\((\w +)(\,\s + (\w +))+\);"; //(completa)
 
-            string dropTable = @"(DROP)\s+(TABLE)\s+(\w+);";
-            string dropDB = @"(DROP)\s+(DATABASE)\s+(\w+);";
-            string createDB = @"(CREATE)\s+(DATABASE)\s+(\w+);";
-            string backupDB = @"(BACKUP\s+DATABASE)\s+(\w+)\s+TO\s+DISK\s*\=\s*\'([^\']+)\';";
+            //Delete
+            string delete = @"DELETE\s+FROM\s+(\w+)\s+WHERE\s+\w+\s*(=|<|>)\s*\w+;";
+
+            //Update
+            string update1 = @"UPDATE\s+(\w+)\s+SET\s+(\w+)\s*\=\s*(\w+)\s+WHERE\s+(\w+)\s*(=|<|>)\s*(\w+);";
+            string update2 = @"UPDATE\s+(\w+)\s+SET\s+(\w+)\s*\=\s*(\w+)(\,\s+(\w+)\s*\=\s*(\w+)\s+)WHERE\s+(\w+)\s*(=|<|>)\s*(\w+);";
+
+            //Drop Table
+            string dropTable = @"DROP\s+TABLE\s+(\w+);";
+
+            //Drop DB
+            string dropDB = @"DROP\s+DATABASE\s+(\w+);";
+
+            //Create DB
+            string createDB = @"CREATE\s+DATABASE\s+(\w+);";
+
+            //Backup DB
+            string backupDB = @"BACKUP\s+DATABASE\s+(\w+)\s+TO\s+DISK\s*\=\s*\'([^\']+)\';";
+
+            //Create Table
             string createTable = @"(CREATE\s+TABLE)\s+(\w+)\s+\((\w+\s+(INT|DOUBLE|TEXT)(\s+|\,\s+\w+\s+(INT|DOUBLE|TEXT))+)\,\s+(PRIMARY\s+KEY)\s+\((\w+)\)\,\s+(FOREIGN\s+KEY)\s+\((\w+)\)\s+REFERENCES\s+(\w+)\s+\((\w+)\);";
 
-            string[] camp1 = null;
-            string camp2;
-            string[] camp3 = null;
-            string campo1;
+            string[] camp1 = new string[10];
+            string camp2 = "";
+            string[] camp3 = new string[10];
+            string campo1 = "";
 
             //SELECT
             Match matchS1 = Regex.Match(query, select1);
@@ -48,8 +66,8 @@ namespace MiniSQLEngine
 
             if (Regex.Match(query, select1).Success)
             {
-                camp1[0] = matchS1.Groups[0].Value;
-                camp2 = matchS1.Groups[1].Value;
+                camp1[0] = matchS1.Groups[1].Value;
+                camp2 = matchS1.Groups[2].Value;
 
                 SQLtype sentencia = new Select(camp2,camp1,camp3);
 
@@ -57,20 +75,22 @@ namespace MiniSQLEngine
             }
             else if (Regex.Match(query, select2).Success)
             {
-                camp1[0] = matchS2.Groups[0].Value;
-                camp2 = matchS2.Groups[1].Value;
-                camp3[0] = matchS2.Groups[2].Value;
-                camp3[1] = matchS2.Groups[2].Value;
-                camp3[2] = matchS2.Groups[2].Value;
-
+                camp1[0] = matchS2.Groups[1].Value;
+                camp2 = matchS2.Groups[2].Value;
+                for(int i = 0; i < matchS2.Groups[3].Length; i++)
+                { 
+                    camp3[i] = matchS2.Groups[3].Captures[i].Value;
+                    
+                }
+                
                 SQLtype sentencia = new Select(camp2, camp1, camp3);
 
                 return sentencia;
             }
             else if (Regex.Match(query, select3).Success)
             {
-                camp1[0] = matchS3.Groups[0].Value;
-                camp2 = matchS3.Groups[1].Value;
+                camp1[0] = matchS3.Groups[1].Value;
+                camp2 = matchS3.Groups[2].Value;
                 
                 SQLtype sentencia = new Select(camp2, camp1, camp3);
 
@@ -78,11 +98,11 @@ namespace MiniSQLEngine
             }
             else if (Regex.Match(query, select4).Success)
             {
-                camp1[0] = matchS4.Groups[0].Value;
-                camp2 = matchS4.Groups[1].Value;
-                camp3[0] = matchS4.Groups[2].Value;
-                camp3[1] = matchS4.Groups[3].Value;
-                camp3[2] = matchS4.Groups[4].Value;
+                camp1[0] = matchS4.Groups[1].Value;
+                camp2 = matchS4.Groups[2].Value;
+                camp3[0] = matchS4.Groups[3].Value;
+                camp3[1] = matchS4.Groups[4].Value;
+                camp3[2] = matchS4.Groups[5].Value;
 
                 SQLtype sentencia = new Select(camp2, camp1, camp3);
 
@@ -90,9 +110,11 @@ namespace MiniSQLEngine
             }
             else if (Regex.Match(query, select5).Success)
             {
-                camp1[0] = matchS5.Groups[0].Value;
+                Console.WriteLine(matchS5.Groups[2].Captures[1].Value);
+                camp1[0] = matchS5.Groups[1].Value;
                 camp1[1] = matchS5.Groups[2].Value;
-                camp2 = matchS5.Groups[3].Value;
+                camp2 = matchS5.Groups[4].Value;
+                
                 
                 SQLtype sentencia = new Select(camp2, camp1, camp3);
 
@@ -114,6 +136,8 @@ namespace MiniSQLEngine
             }
 
 
+            //Drop table
+
             Match matchDT = Regex.Match(query, dropTable);
 
             if (matchDT.Success)
@@ -125,8 +149,34 @@ namespace MiniSQLEngine
                 return sentencia;
             }
 
+
+            string cinsert1;
+            string[] cinsert2 = null;
+            string[] cinsert3 = null;
+
+            //Insert 
+
+            Match matchI1 = Regex.Match(query, insert1);
+            Match matchI2 = Regex.Match(query, insert2);
+            Match matchI3 = Regex.Match(query, insert3);
+            Match matchI4 = Regex.Match(query, insert4);
+
+            if (Regex.Match(query, insert1).Success)
+            {
+                cinsert1 = matchI1.Groups[1].Value;
+                cinsert3[0] = matchI1.Groups[2].Value;
+
+                SQLtype sentencia = new Select(cinsert1, cinsert2, cinsert3);
+
+                return sentencia;
+            }
+           
+
+
+
+
             Match match3 = Regex.Match(query, delete);
-            Match match4 = Regex.Match(query, update);
+            //Match match4 = Regex.Match(query, update);
             Match match5 = Regex.Match(query, dropTable);
             Match match6 = Regex.Match(query, dropDB);
             Match match7 = Regex.Match(query, createDB);
@@ -134,7 +184,7 @@ namespace MiniSQLEngine
             Match match9 = Regex.Match(query, createTable);
 
 
-
+            return null;
 
         }
